Fixed formatting issues and moved formatter test
