+++ turbinia/api/client/.gitlab-ci.yml
@@ -6,9 +6,9 @@ stages:
 .tests:
   stage: test
   script:
-    - pip install -r requirements.txt
-    - pip install -r test-requirements.txt
-    - pytest --cov=turbinia_api_lib
+   - pip install -r requirements.txt
+   - pip install -r test-requirements.txt
+   - pytest --cov=turbinia_api_lib
 
 test-3.6:
   extends: .tests
