+++ turbinia/state_manager.py
@@ -30,8 +30,8 @@ from datetime import timedelta
 import six
 
 from turbinia import config
-from turbinia import TurbiniaException
 from turbinia.config import DATETIME_FORMAT
+from turbinia import TurbiniaException
 
 config.LoadConfig()
 if 'unittest' in sys.modules.keys():
